+++ FinTrac/ProjectTests/Dtos Tests/MovementInXDays_Tests.cs
@@ -24,7 +24,7 @@ namespace BusinessLogicTests.Dto_Components
         [TestMethod]
         public void GivenSpendingArray_ShouldBeSet()
         {
-            int[] spendings = new int[31];
+            decimal[] spendings = new decimal[31];
             MovementInXDaysDTO movements = new MovementInXDaysDTO();
             movements.Spendings = spendings;
